refactor: share the Cloud app-selection resolution between the two commands
