+++ src/lib/cloud/cloud-remove-sheet-icons.js
@@ -13,8 +13,7 @@ import {
     SHEET_SKIPPED,
 } from '../util/sheet-list.js';
 import { withEngineSession } from '../util/engine-session.js';
-import { listAppsByCollection } from './cloud-apps.js';
-import { toAppIdList } from '../util/app-ids.js';
+import { resolveCloudAppSelection } from './cloud-app-selection.js';
 import { CLEAR_REASON } from '../util/sheet-decision-reasons.js';
 import {
     runOverAppsWithReport,
@@ -300,8 +299,6 @@ export const qscloudRemoveSheetIcons = async (options) => {
         logger.debug(`BSI executable path: ${bsiExecutablePath}`);
         logger.debug(`Options: ${JSON.stringify(redactOptions(options), null, 2)}`);
 
-        const appIdsToProcess = [];
-
         // Get array of all available collections
         const cloudConfig = {
             url: options.tenanturl,
@@ -330,31 +327,14 @@ export const qscloudRemoveSheetIcons = async (options) => {
             return false;
         }
 
-        // Apps named directly. --appid is variadic, so this is a list.
-        const namedAppIds = toAppIdList(options.appid);
-        appIdsToProcess.push(...namedAppIds);
-
-        // --appid and --collectionid are additive, not alternatives: apps named either
-        // way are all processed. runOverApps() dedupes, so an app that is both named by
-        // --appid and in the collection is still processed once.
-        let collectionAppIds = [];
-        const useCollection = Boolean(options.collectionid && options.collectionid.length > 0);
-        if (useCollection) {
-            const apps = await listAppsByCollection(saasInstance, options.collectionid);
-            logger.verbose(`Collection '${options.collectionid}' exists`);
-            collectionAppIds = apps.map((app) => app.id);
-            appIdsToProcess.push(...collectionAppIds);
-        }
+        // Selection resolution is shared with the other Cloud command; the
+        // provenance it returns feeds the run report directly.
+        const selection = await resolveCloudAppSelection(saasInstance, options);
 
         return await runOverAppsWithReport({
             command: 'qscloud remove-sheet-icons',
             dryRun,
-            appIds: appIdsToProcess,
-            namedAppIds,
-            selectorAppIds: collectionAppIds,
-            selector: useCollection
-                ? { option: 'collectionid', value: options.collectionid }
-                : null,
+            ...selection,
             logPrefix: { plan: 'CLOUD PLAN REMOVE ICONS', process: 'CLOUD REMOVE SHEET ICONS' },
             emptySelectionHint: 'Check the --appid and --collectionid options.',
             planApp: (appId, report) =>
